BE/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.4
defusedxml==0.8.0rc2
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.11
gunicorn==22.0.0
django-admin-interface==0.26.0
django-colorfield==0.10.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-heroku==0.3.1
django-import-export==3.2.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.26.5
et-xmlfile==1.1.0
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
MarkupPy==1.14
MarkupSafe==2.1.3
oauthlib==3.2.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.1.2
Pillow==10.3.0
pycparser==2.21
PyJWT==2.8.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.32.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
simplejson==3.19.2
social-auth-app-django==5.4.1
social-auth-core==4.5.0
sqlparse==0.5.0
tablib==3.5.0
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
pandas==2.2.2
